media-source/SourceBuffer-appendWindowStart_invalid.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#appendWindowStart with new invalid value</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        var sourceBuffer = mediaSource.addSourceBuffer(mime);
        t.step(function() {
            assert_throws({name: 'InvalidAccessError'},
                          function() {
                              sourceBuffer.appendWindowStart = -1;
                          },
                          'The new value is less than 0');
            assert_throws({name: 'InvalidAccessError'},
                          function() {
                              sourceBuffer.appendWindowStart = 0;
                              sourceBuffer.appendWindowEnd = 1;
                              sourceBuffer.appendWindowStart = sourceBuffer.appendWindowEnd + 1;
                          },
                          'The new value is greater than appendWindowEnd');
            assert_throws({name: 'InvalidAccessError'},
                          function() {
                              sourceBuffer.appendWindowStart = 0;
                              sourceBuffer.appendWindowEnd = 1;
                              sourceBuffer.appendWindowStart = sourceBuffer.appendWindowEnd;
                          },
                          'The new value is equal to appendWindowEnd');
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
},
'SourceBuffer#appendWindowStart : If the new value is less than 0 or greater than or equal to appendWindowEnd, INVALID_ACCESS_ERR exception should be thrown.');
</script>
</body>
</html>
